move bridge server connection to new file, post that the cli is building in copilotconnect
