+++ .github/workflows/release.yml
@@ -19,6 +19,19 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
 
+      - name: Get PR Description
+        id: pr_info
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          echo -e '${{ github.event.pull_request.body }}' > "notes.txt"
+          if [ -z "$PR_BODY" ]; then
+            PR_BODY="No description provided"
+          fi
+          echo "description<<EOF" >> $GITHUB_OUTPUT
+          echo "$PR_BODY" >> $GITHUB_OUTPUT
+          echo "EOF" >> $GITHUB_OUTPUT
+
       - name: Fetch Package Version
         id: version
         run: |
@@ -33,7 +46,7 @@ jobs:
             git push origin "v$VERSION"
             echo "skip=false" >> $GITHUB_OUTPUT
           fi
-          
+
       # Create initial pre-release
       - name: Create Pre-Release
         if: steps.version.outputs.skip == 'false'
@@ -42,9 +55,7 @@ jobs:
         run: |
           gh release create ${{ steps.version.outputs.version }} \
             --title "Release ${{ steps.version.outputs.version }}" \
-            --notes "Automated release from ${{ github.ref_name }} branch
-
-                    ${{ steps.pr_info.outputs.description }}" \
+            --notes-file notes.txt \
             --target ${{ github.sha }} \
             --prerelease
 
@@ -54,7 +65,7 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: gh release upload ${{ steps.version.outputs.version }} ./release/*
-        
+
       # Convert pre-release to latest release
       - name: Mark as Latest Release
         if: steps.version.outputs.skip == 'false'
