+++ Windows/Lib/DataClasses/SSoTmeProject.cs
@@ -699,263 +699,6 @@ namespace SSoTme.OST.Lib.DataClasses
             }
         }
         
-        private (bool, string, string) GetLastCopilotRequestedRead(string readReqUri, string userid)
-        {
-            using (var httpClient = new HttpClient())
-            {
-                httpClient.DefaultRequestHeaders.Add("X-Microsoft-TenantID", userid);
-                string response = httpClient.GetStringAsync(readReqUri).Result;
-                var json = JsonDocument.Parse(response);
-                var changedVal = json.RootElement.GetProperty("changed").GetRawText();
-                bool changed = changedVal == "true";
-                
-                string timestamp = "0";
-                if (changed) {
-                    timestamp = json.RootElement.GetProperty("timestamp").GetString();
-                    if (string.IsNullOrEmpty(timestamp)) { Console.WriteLine($"WARN: Received null timestamp in response: {response}"); }
-                }
-                return (changed, response, timestamp);
-            }
-        }
-
-        private string PostDataToBridge(JToken data, string uri, string dataTimestamp, string userid)
-        {
-            using (var httpClient = new HttpClient())
-            {
-                httpClient.DefaultRequestHeaders.Add("X-Microsoft-TenantID", userid);
-                try
-                {
-                    // Add timestamp to the data
-                    data["timestamp"] = dataTimestamp;
-                    
-                    var jsonString = data.ToString(Newtonsoft.Json.Formatting.None);
-                    // Console.WriteLine($"Posting to bridge: {uri} - {data}");
-                    var content = new StringContent(jsonString, Encoding.UTF8, "application/json");
-                    var response = httpClient.PostAsync(uri, content).Result;
-
-                    if (response.IsSuccessStatusCode)
-                    {
-                        return response.Content.ReadAsStringAsync().Result;
-                    }
-                    else
-                    {
-                        this.LogMessage("Error posting data to bridge: {0} - {1}", response.StatusCode, response.ReasonPhrase);
-                        return null;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    this.LogMessage("Exception posting data to bridge: {0}", ex.Message);
-                    return null;
-                }
-            }
-        }
-        
-        private (JToken, bool) RunCopilotAction(string commandData, string baseId, BaserowBackend baserowClient)
-        {  // todo you can make undo/redo actions using the baserow 'ClientSessionId' header
-            
-            var validActions = new[] { "list_tables", "update_cell", "get_cell", "get_table_fields", "create_row", "create_field", "update_field", "delete_field" };
-            try
-            {
-                var requestedChanges = JsonConvert.DeserializeObject<dynamic>(commandData);
-                
-                Console.WriteLine($"{requestedChanges.reason} (action: '{requestedChanges.action}')"); // copilot will provide this with every request
-                
-                // match copilot's requested action to the right baserow endpoint
-                string tableId = requestedChanges.tableId;
-                if ((requestedChanges.action != "list_tables" && requestedChanges.action != "update_field" && requestedChanges.action != "delete_field") && requestedChanges.tableId == null) {
-                    return (new JObject
-                    {
-                        ["content"] = null,
-                        ["msg"] = "Error applying changes: This endpoint requires tableId!"
-                    }, false);
-                }
-                
-                string fieldId = requestedChanges.fieldId;
-                if (requestedChanges.fieldId == null && (requestedChanges.action == "update_field" || requestedChanges.action == "delete_field"))
-                {
-                    return (new JObject
-                    {
-                        ["content"] = null,
-                        ["msg"] = "Error applying changes: this endpoint requires fieldId!"
-                    }, false);
-                }
-                
-                string rowId = requestedChanges.rowId;
-                if (requestedChanges.action == "move_row" && requestedChanges.rowId == null)
-                {
-                    return (new JObject
-                    {
-                        ["content"] = null,
-                        ["msg"] = "Error applying changes: this endpoint requires rowId!"
-                    }, false);
-                }
-                else if ((requestedChanges.action == "update_cell" || requestedChanges.action == "get_cell") &&
-                    (requestedChanges.rowId == null || requestedChanges.fieldId == null)) 
-                {
-                    return (new JObject
-                    {
-                        ["content"] = null,
-                        ["msg"] = "Error applying changes: this endpoint requires rowId and fieldId!"
-                    }, false);
-                }
-
-                // MARK START action definitions
-                if (requestedChanges.action == "list_tables")
-                {
-                    // table ids are globally unique across all databases
-                    // this.LogMessage("Fetching tables for base: {0}", baseId);
-                    var tablesData = baserowClient.FetchTablesForBase(baseId);
-                    return (new JObject
-                    {
-                        ["content"] = tablesData,
-                        ["msg"] = $"Successfully retrieved tables for base: {baseId}"
-                    }, false);
-                }
-                else if (requestedChanges.action == "get_table_fields")
-                {
-                    // Console.WriteLine($"Fetching table data with field mappings for tableId: {tableId}");
-                    JToken tableSchema = baserowClient.GetTableSchema(tableId);
-                    return (new JObject
-                    {
-                        ["content"] = tableSchema,
-                        ["msg"] = $"Successfully retrieved table fields for: {tableId}"
-                    }, false);
-                }
-                else if (requestedChanges.action == "create_field")
-                {
-                    string name = requestedChanges.content.fieldName;
-                    string type = requestedChanges.content.fieldType;
-                    Console.WriteLine($"Creating new field: name: {name}, type: {type}");
-                    JToken resp = baserowClient.CreateField(tableId, name, type);
-                    return (new JObject
-                    {
-                        ["content"] = resp,
-                        ["msg"] = $"Successfully created field on tableId: {tableId}"
-                    }, true);
-                }
-                else if (requestedChanges.action == "update_field")
-                {
-                    string name = requestedChanges.content.fieldName;
-                    string type = requestedChanges.content.fieldType;
-                    // Console.WriteLine($"Updating field: id: {fieldId}, to type: {type}, name: {name}");
-                    JToken resp = baserowClient.UpdateField(fieldId, name, type);
-                    return (new JObject
-                    {
-                        ["content"] = resp,
-                        ["msg"] = $"Successfully updated field: {fieldId}"
-                    }, true);
-                }
-                else if (requestedChanges.action == "delete_field")
-                {
-                    Console.WriteLine($"Deleting field: id: {fieldId}");
-                    JToken resp = baserowClient.DeleteField(fieldId);
-                    return (new JObject
-                    {
-                        ["content"] = resp,
-                        ["msg"] = $"Successfully deleted field: {fieldId}"
-                    }, true);
-                }
-                else if (requestedChanges.action == "create_row")
-                {
-                    Console.WriteLine($"Creating new row");
-                    JToken resp = baserowClient.CreateRow(tableId);
-                    return (new JObject
-                    {
-                        ["content"] = resp,
-                        ["msg"] = $"Successfully created row"
-                    }, true);
-                }
-                else if (requestedChanges.action == "move_row")
-                {
-                    JToken resp;
-                    if (requestedChanges.ContainsKey("relativeRowId"))
-                    {
-                        resp = baserowClient.MoveRow(tableId, rowId, requestedChanges.relativeRowId);
-                        Console.WriteLine($"Moved row {rowId} to position before {requestedChanges.relativeRowId}");
-                        return (new JObject
-                        {
-                            ["content"] = resp,
-                            ["msg"] = $"Successfully moved row {rowId} to position before {requestedChanges.relativeRowId}"
-                        }, true);
-                    }
-                    resp = baserowClient.MoveRow(tableId, rowId);
-                    Console.WriteLine($"Moved row {rowId} to the end of the table");
-                    return (new JObject
-                    {
-                        ["content"] = resp,
-                        ["msg"] = $"Successfully moved row {rowId} to the end of the table"
-                    }, true);
-                }
-                else if (requestedChanges.action == "get_cell")
-                {
-                    // Console.WriteLine($"Fetching cell data for field x row: {fieldId}x{rowId}");
-                    JToken fieldResp = baserowClient.GetCell(tableId, rowId, fieldId);
-                    return (new JObject
-                    {
-                        ["content"] = fieldResp,
-                        ["msg"] = $"Successfully retrieved cell: {fieldId}x{rowId}"
-                    }, false);
-                }
-                else if (requestedChanges.action == "update_cell")
-                {
-                    string newValue = requestedChanges.content;
-                    // Console.WriteLine($"Updating cell data for field x row: {fieldId}x{rowId} to {newValue}");
-                    JToken resp = baserowClient.UpdateCell(tableId, rowId, fieldId, newValue);
-                    return (new JObject
-                    {
-                        ["content"] = resp,
-                        ["msg"] = $"Successfully updated cell contents: {fieldId}x{rowId}"
-                    }, true);
-                }
-                
-                return (new JObject
-                {
-                    ["content"] = null,
-                    ["msg"] = $"Action was not matched to any of the following for base {baseId}: {string.Join(", ", validActions)}"
-                }, false);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error running copilot action: {ex.Message}");
-                return (new JObject
-                {
-                    ["content"] = null,
-                    ["msg"] = $"Error running copilot action: {ex.Message}"
-                }, false);
-            }
-        }
-
-        private bool BaseIsIn(string baseId, JToken userBases)
-        {
-            if (string.IsNullOrEmpty(baseId) || userBases == null)
-            {
-                return false;
-            }
-
-            // userBases is expected to be a JArray or a JToken containing a "results" array
-            JArray basesArray = userBases.Type == JTokenType.Array
-                ? (JArray)userBases
-                : userBases["results"] as JArray;
-
-            if (basesArray == null)
-            {
-                Console.WriteLine("Could not find a list of bases in provided JToken.");
-                return false;
-            }
-
-            foreach (var baseToken in basesArray)
-            {
-                var id = baseToken["id"]?.ToString();
-                if (!string.IsNullOrEmpty(id) && id == baseId)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-        
         private void ListenForChangesAndRebuild(
     string buildPath,
     bool includeDisabled,
@@ -969,36 +712,37 @@ namespace SSoTme.OST.Lib.DataClasses
             bool changeEverDetected = false;
             string baseUri = $"https://ssotme-cli-airtable-bridge-ahrnz660db6k4.aws-us-east-1.controlplane.us";
             string baseCopilotUri = "https://ssotme-cli-airtable-bridge-v2-ahrnz660db6k4.cpln.app/copilot";
-            string copilotReadUri = $"{baseCopilotUri}/check-req-actions?baseId={baseId}";
             Console.WriteLine($"Polling {baseUri}/check?baseId={baseId} for changes to base: `{baseId}`...");
 
-            BaserowBackend baserowClient = new SSoTme.OST.Core.Lib.External.BaserowBackend();
-            string microsoftTenantUserId = String.Empty;
+            string microsoftTenantUserId = "test";  // todo actually use the user's microsoft account tenant id somehow
+            BOTBridgeServerConnector conn = new BOTBridgeServerConnector(microsoftTenantUserId, baseCopilotUri, baseId);
+            
             if (isCopilot)
             {
-                microsoftTenantUserId = "test";  // todo actually use the user's microsoft account tenant id somehow
-                Console.WriteLine($"Polling {copilotReadUri} for copilot actions...");
-                // get baserow client from ~/.ssotme/ssotme.key file -> "baserow" api
-                baserowClient.InitFromHomeFile();
-                JToken userBaserowBases = PostAvailableBases($"{baseCopilotUri}/available-bases", baserowClient, microsoftTenantUserId);
+                Console.WriteLine($"Polling {conn.copilotReadUri} for copilot actions...");
+                JToken userBaserowBases = conn.PostAvailableBases();
                 if (userBaserowBases == null)
                 {
                     throw new Exception("Couldn't post to the remote server");
                 }
-                else if (!BaseIsIn(baseId, userBaserowBases))
+                else if (!conn.CurrentBaseIdIsAvailable(userBaserowBases))
                 {
                     throw new Exception($"You don't have access to any baserow database with ID: {baseId}");
                 }
             }
             while (true) {
                 if (isCopilot) {
-                    var (newCopilotActionRequest, copilotProvidedData, timestamp) = GetLastCopilotRequestedRead(copilotReadUri, microsoftTenantUserId);
+                    var (newCopilotActionRequest, copilotProvidedData, timestamp) = conn.GetLastCopilotRequestedRead();
                     if (newCopilotActionRequest) {
                         if (!string.IsNullOrEmpty(copilotProvidedData))
                         {
-                            var (response, contentWasUpdated) = RunCopilotAction(copilotProvidedData, baseId, baserowClient);
-                            PostDataToBridge(response, $"{baseCopilotUri}/put-action-result?baseId={baseId}", timestamp, microsoftTenantUserId);
-                            if (contentWasUpdated) PostChange(baseUri, baseId);  // signal a rebuild is necessary
+                            var (response, contentWasUpdated) = conn.RunCopilotAction(copilotProvidedData);
+                            conn.PostDataToBridge(response, timestamp);
+                            if (contentWasUpdated)
+                            {
+                                conn.PostChange(baseUri);  // signal a rebuild is necessary
+                                conn.SetIsBuilding(true);
+                            }
                         }
                     }
                 }
@@ -1016,7 +760,17 @@ namespace SSoTme.OST.Lib.DataClasses
                     if (elapsed > 10)
                     {
                         Console.WriteLine("No changes in last 10 seconds. Rebuilding...");
-                        this.DoRebuild(buildPath, includeDisabled, transpilerGroup, isBuildLocal, isBuildAll);
+                        try
+                        {
+                            this.DoRebuild(buildPath, includeDisabled, transpilerGroup, isBuildLocal, isBuildAll);
+                        }
+                        catch (Exception ex)
+                        {
+                            Console.WriteLine($"Rebuild failed; {ex.Message}");
+                        }
+                        
+                        if (isCopilot) conn.SetIsBuilding(false);
+
                         changeEverDetected = false;
                         lastChangedTime = null;
                     }
@@ -1050,71 +804,6 @@ namespace SSoTme.OST.Lib.DataClasses
             {
                 Console.WriteLine($"Error calling listener API: {ex.Message}");
                 return false;
-            }
-        }
-
-        private void PostChange(string uri, string baseId)
-        {
-            try
-            {
-                using (var httpClient = new HttpClient())
-                {
-                    var response = httpClient.GetAsync($"{uri}/mark?baseId={baseId}").Result;
-
-                    if (!response.IsSuccessStatusCode)
-                    {
-                        var errorContent = response.Content.ReadAsStringAsync().Result;
-                        Console.WriteLine($"Failed to post change. Status: {response.StatusCode}, Error: {errorContent}");
-                    }
-                    else
-                    {
-                        var result = response.Content.ReadAsStringAsync().Result;
-                        Console.WriteLine($"Change posted successfully: {result}");
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error posting change to listener API: {ex.Message}");
-            }
-        }
-
-        private JToken PostAvailableBases(string uri, BaserowBackend baserowClient, string userid)
-        {
-            JToken availableBases = baserowClient.GetAvailableBases();
-
-            try
-            {
-                using (var httpClient = new HttpClient())
-                {
-                    httpClient.DefaultRequestHeaders.Add("X-Microsoft-TenantID", userid);
-                    var payload = new JObject
-                    {
-                        ["bases"] = availableBases
-                    };
-
-                    var json = JsonConvert.SerializeObject(payload);
-                    var content = new StringContent(json, Encoding.UTF8, "application/json");
-                    Console.WriteLine($"Posting available bases: {availableBases.ToString(Formatting.None)} for user: {userid}");
-                    var response = httpClient.PostAsync(uri, content).Result;
-
-                    if (!response.IsSuccessStatusCode)
-                    {
-                        var errorContent = response.Content.ReadAsStringAsync().Result;
-                        Console.WriteLine($"Failed to post bases. Status: {response.StatusCode}, Error: {errorContent}");
-                        return null;
-                    }
-                    else
-                    {
-                        var result = response.Content.ReadAsStringAsync().Result;
-                        return availableBases;
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error posting available user bases to Bridge API: {ex.Message}");
-                return null;
             }
         }
