+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ssotme",
-  "version": "2025.06.16",
+  "version": "2025.06.17",
   "dotnet": "8.0.410",
   "description": "",
   "bin": {
